move test specs to a tab structure
