+++ web/src/components/TestOutputs/TestOutputs.tsx
@@ -1,0 +1,5 @@
+const TestOutputs = () => {
+  return <div>Hola</div>;
+};
+
+export default TestOutputs;
