+++ web/src/components/TestOutputs/index.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-restricted-exports
+export {default} from './TestOutputs';
