+++ web/src/components/TestResults/TestResults.styled.ts
@@ -3,7 +3,6 @@ import styled from 'styled-components';
 
 export const Container = styled.div`
   height: 100%;
-  padding: 24px;
   position: relative;
 `;
