+++ web/src/components/TestResults/TestResults.tsx
@@ -1,8 +1,6 @@
 import {useCallback} from 'react';
 
 import LoadingSpinner from 'components/LoadingSpinner';
-import TestSpecDetail from 'components/TestSpecDetail';
-import {useTestSpecForm} from 'components/TestSpecForm/TestSpecForm.provider';
 import TestSpecs from 'components/TestSpecs';
 import {useSpan} from 'providers/Span/Span.provider';
 import {useTestSpecs} from 'providers/TestSpecs/TestSpecs.provider';
@@ -12,11 +10,15 @@ import AssertionAnalyticsService from 'services/Analytics/AssertionAnalytics.ser
 import {TAssertionResultEntry} from 'types/Assertion.types';
 import Header from './Header';
 import * as S from './TestResults.styled';
-import AssertionService from '../../services/Assertion.service';
 
-const TestResults = () => {
-  const {open} = useTestSpecForm();
-  const {isLoading, assertionResults, remove, revert, setSelectedSpec, selectedTestSpec} = useTestSpecs();
+interface IProps {
+  onDelete(selector: string): void;
+  onEdit(assertionResult: TAssertionResultEntry): void;
+  onRevert(originalSelector: string): void;
+}
+
+const TestResults = ({onDelete, onEdit, onRevert}: IProps) => {
+  const {isLoading, assertionResults, setSelectedSpec} = useTestSpecs();
   const {selectedSpan, onSetFocusedSpan, onSelectSpan} = useSpan();
   const {totalFailedSpecs, totalPassedSpecs} = useAppSelector(TestSpecsSelectors.selectTotalSpecs);
 
@@ -32,51 +34,6 @@ const TestResults = () => {
     [assertionResults?.resultList, onSelectSpan, onSetFocusedSpan, setSelectedSpec]
   );
 
-  const handleClose = useCallback(() => {
-    onSetFocusedSpan('');
-    setSelectedSpec();
-  }, [onSetFocusedSpan, setSelectedSpec]);
-
-  const handleEdit = useCallback(
-    ({selector, resultList: list}: TAssertionResultEntry) => {
-      AssertionAnalyticsService.onAssertionEdit();
-
-      open({
-        isEditing: true,
-        selector,
-        defaultValues: {
-          assertions: list.map(({assertion}) => AssertionService.getStructuredAssertion(assertion)),
-          selector,
-        },
-      });
-    },
-    [open]
-  );
-
-  const handleDelete = useCallback(
-    (selector: string) => {
-      AssertionAnalyticsService.onAssertionDelete();
-      remove(selector);
-    },
-    [remove]
-  );
-
-  const handleRevert = useCallback(
-    (originalSelector: string) => {
-      AssertionAnalyticsService.onRevertAssertion();
-      revert(originalSelector);
-    },
-    [revert]
-  );
-
-  const handleSelectSpan = useCallback(
-    (spanId: string) => {
-      onSelectSpan(spanId);
-      onSetFocusedSpan(spanId);
-    },
-    [onSelectSpan, onSetFocusedSpan]
-  );
-
   return (
     <S.Container>
       <Header selectedSpan={selectedSpan!} totalFailedSpecs={totalFailedSpecs} totalPassedSpecs={totalPassedSpecs} />
@@ -90,23 +47,12 @@ const TestResults = () => {
       {!isLoading && Boolean(assertionResults) && (
         <TestSpecs
           assertionResults={assertionResults!}
-          onDelete={handleDelete}
-          onEdit={handleEdit}
+          onDelete={onDelete}
+          onEdit={onEdit}
           onOpen={handleOpen}
-          onRevert={handleRevert}
+          onRevert={onRevert}
         />
       )}
-
-      <TestSpecDetail
-        isOpen={Boolean(selectedTestSpec)}
-        onClose={handleClose}
-        onDelete={handleDelete}
-        onEdit={handleEdit}
-        onRevert={handleRevert}
-        onSelectSpan={handleSelectSpan}
-        selectedSpan={selectedSpan?.id}
-        testSpec={selectedTestSpec}
-      />
     </S.Container>
   );
 };
